write files to master dir on blob store
Reason:
Earlier, the main branch was "master" and files were written to the branch name.
Now, the main branch is "main".
But the frontend expects the app store data under the "master" dir.
Instead of changing the frontend - which would be the correct way - I am writing to the "master" dir even though the branch is called "main".
Changing the frontend would be the correct move though.