write files to both master and the branch name on blob store
Reason:
We want to change to using "main" as the main branch for good which includes changing the freeshard frontend to query the app store from the "main" directory. Currently, it queries from "master".
To make that transition easy, we keep both app stores current, so it does not matter when the change in the frontend will occur. The behaviour will be the same.